Add key-binding for executing scripts
The python interpreter is the same active the working shell (the one
on top which Vim is executed). Therefore, vim is naturally
integrated with conda.
In order for the binding to work properly, one should start vim
*after* the activation of a conda environment.
Otherwise, one can put Vim in background with <C-z>, activate the
environment and launch the execution.

Bindings are added for both NORMAL and INSERT mode.