Change statusline
Statusline has been modified to fix a problem encountered with
visual-block mode. The `mode()` built-in function could not detect
properly the mode state when triggering visual-block mode.

To fix it a new function has been added that puts the output of
`mode()` to visual-block mode if it is not present in the
currentmode dictionary.

Other minor and aesthetics improvements are been added as well.

TODO:
- add custom colors for each mode