Add file type-specific tabular mappings
All the mappings are triggered as follows:
<Leader>a<comment_string>

So for Python it is <Leader>a#, for C++ it is <Leader>a/, etc.
This allows to align comment and improve the readability of code
very rapidly.
The align autoload function has been removed from vimrc but I
decided to keep it in the autoload folder for future tweaking.