Add filetype marking
The major change in the vimrc file is the filetype marker. The last
file of the most used filetypes (.py, .cpp, ...) are marked such
that we can easily recall them in another editing session.
The markers are the following:
*.cpp   C
*.py    P
*.md    M
*.txt   T
*vimrc  V

So using `<marker> we immediately open the last file edited of the
given extension.

Other minor changes has been implemented too. The `K` key has been
mapped to `<Nop>` to remove the annoying behaviour. Moreover, the
resizing mapping have been updated to resize the width of the
windows.
A new mapping (Leader>0 <C-w>=) has been added to make all split in the
current window equal.